# src/main/resources/openapi/finscope-api.yaml
openapi: 3.0.3
info:
  title: FinScope API
  description: |
    API reactiva de FinScope para la gestion de finanzas personales.

    El modelo se reduce a cuatro piezas: la transaccion, su tipo, su categoria y sus
    tags.
    El tipo es un catalogo cerrado de dos valores, ingreso y egreso, porque de el
    depende el signo del importe.

    La categoria y los tags responden preguntas distintas y no son intercambiables. La
    categoria dice en que se gasto y es exactamente una por transaccion, de modo que la
    suma por categoria coincide con el total del periodo y puede repartirse en
    porcentajes. Los tags dicen con quien, para que o en que contexto ocurrio el gasto,
    son cero o varios por transaccion y por tanto se solapan: sus importes nunca reparten
    un total. Ambos son catalogos del usuario, editables y compartidos entre sus
    transacciones.

    Todos los recursos financieros pertenecen al usuario autenticado y estan aislados
    entre usuarios. Salvo los endpoints de autenticacion, todas las operaciones exigen
    un token de acceso.
  version: 6.0.0
servers:
  - url: /
    description: Servidor local
security:
  - bearerAuth: []
...
paths:
...
  /categories:
    get:
      tags: [categories]
      summary: Lista las categorias del usuario
      description: |
        Devuelve el catalogo completo de categorias del usuario, ordenado alfabeticamente,
        con el numero de transacciones que clasifica cada una.
        El catalogo se siembra al registrarse con un juego inicial de categorias y desde
        ahi es del usuario: puede anadir, renombrar y eliminar las que quiera.
      operationId: listCategories
      responses:
        "200":
          description: Listado de categorias
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryResponse"
    post:
      tags: [categories]
      summary: Crea una categoria
      description: |
        El nombre debe estar libre dentro del catalogo del usuario sin distinguir
        mayusculas, igual que en los tags.
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveCategoryRequest"
      responses:
        "201":
          description: Categoria creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /categories/{id}:
    parameters:
      - $ref: "#/components/parameters/PathId"
    patch:
      tags: [categories]
      summary: Actualiza una categoria
      description: |
        Cambia el nombre o el tipo de movimiento al que se ofrece la categoria. El cambio
        alcanza a todas las transacciones que clasifica, porque la categoria es una sola y
        se comparte entre ellas.
        Los campos ausentes se dejan como estan. Renombrar a un nombre ya ocupado se
        rechaza en lugar de fusionar las dos categorias, para que la operacion nunca
        destruya informacion sin avisar.
      operationId: updateCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveCategoryRequest"
      responses:
        "200":
          description: Categoria actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [categories]
      summary: Elimina una categoria
      description: |
        Elimina la categoria del catalogo. Las transacciones que clasificaba pasan a la
        categoria de reserva del usuario, la marcada con isSystem: como toda transaccion
        tiene exactamente una categoria, no pueden quedarse sin ninguna.
        La categoria de reserva no se puede eliminar.
      operationId: deleteCategory
      responses:
        "204":
          description: Categoria eliminada
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /tags:
    get:
      tags: [tags]
      summary: Lista los tags del usuario
      description: |
        Devuelve el catalogo completo de tags del usuario, ordenado alfabeticamente, con el
        numero de transacciones que lleva cada uno. Sirve para ofrecer autocompletado al
        escribir y para administrar el catalogo.
        Se incluyen los tags que todavia no usa ninguna transaccion, con transactionCount a
        cero: siguen ocupando su nombre, asi que el usuario debe poder verlos y borrarlos.
      operationId: listTags
      responses:
        "200":
          description: Listado de tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TagResponse"
    post:
      tags: [tags]
      summary: Crea un tag en el catalogo del usuario
      description: |
        Alta explicita de un tag, para preparar el catalogo sin registrar una transaccion.
        No es la via habitual: los tags siguen naciendo al escribirlos dentro de una.
      operationId: createTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveTagRequest"
      responses:
        "201":
          description: Tag creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /tags/{id}:
    parameters:
      - $ref: "#/components/parameters/PathId"
    patch:
      tags: [tags]
      summary: Renombra un tag
      description: |
        Cambia el nombre del tag en todas las transacciones que lo llevan, porque el tag es
        uno solo y se comparte entre ellas.
        El nombre debe seguir libre dentro del catalogo del usuario sin distinguir
        mayusculas: renombrar a un nombre ya existente se rechaza en lugar de fusionar los
        dos tags, para que la operacion nunca destruya informacion sin avisar.
      operationId: updateTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveTagRequest"
      responses:
        "200":
          description: Tag renombrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [tags]
      summary: Elimina un tag
      description: |
        Elimina el tag del catalogo y lo retira de todas las transacciones que lo llevan.
        Las transacciones no se tocan: solo dejan de estar clasificadas por el.
      operationId: deleteTag
      responses:
        "204":
          description: Tag eliminado
        "404":
          $ref: "#/components/responses/NotFound"
  /transactions:
    get:
      tags: [transactions]
      summary: Lista transacciones filtradas y paginadas
      description: |
        Todos los filtros son opcionales y se combinan con AND.
        Los parametros month y year son un atajo para acotar a un mes natural completo
        y no pueden combinarse con dateFrom o dateTo.
      operationId: listTransactions
      parameters:
        - $ref: "#/components/parameters/MonthFilter"
        - $ref: "#/components/parameters/YearFilter"
        - $ref: "#/components/parameters/DateFromFilter"
        - $ref: "#/components/parameters/DateToFilter"
        - $ref: "#/components/parameters/TransactionTypeIdFilter"
        - $ref: "#/components/parameters/CategoryIdFilter"
        - $ref: "#/components/parameters/TagFilter"
        - name: page
          in: query
          description: Numero de pagina, empezando en cero
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Cantidad de elementos por pagina
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: sort
          in: query
          description: |
            Criterio de ordenamiento con formato campo,direccion.
            Campos admitidos date, amount e id. Direcciones asc y desc.
          required: false
          schema:
            type: string
            default: date,desc
      responses:
        "200":
          description: Pagina de transacciones
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionPageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      tags: [transactions]
      summary: Crea una transaccion
      operationId: createTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionRequest"
      responses:
        "201":
          description: Transaccion creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /transactions/{id}:
    parameters:
      - $ref: "#/components/parameters/PathId"
    get:
      tags: [transactions]
      summary: Obtiene una transaccion por su identificador
      operationId: getTransactionById
      responses:
        "200":
          description: Transaccion encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [transactions]
      summary: Actualiza parcialmente una transaccion
      description: |
        Solo se modifican los campos presentes en la peticion.
        Si tags esta presente reemplaza por completo los tags de la transaccion;
        si se omite, los tags actuales se conservan.
      operationId: updateTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTransactionRequest"
      responses:
        "200":
          description: Transaccion actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [transactions]
      summary: Elimina una transaccion
      description: |
        La transaccion se elimina fisicamente junto con sus tags, que no existen fuera
        de ella.
      operationId: deleteTransaction
      responses:
        "204":
          description: Transaccion eliminada
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token de acceso emitido por /auth/login o /auth/refresh
  parameters:
    PathId:
      name: id
      in: path
      description: Identificador del recurso
      required: true
      schema:
        type: integer
        format: int64
    MonthFilter:
      name: month
      in: query
      description: Mes de la transaccion, entre 1 y 12
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 12
    YearFilter:
      name: year
      in: query
      description: Anio de la transaccion
      required: false
      schema:
        type: integer
        minimum: 1970
        maximum: 9999
    DateFromFilter:
      name: dateFrom
      in: query
      description: Fecha inicial inclusiva
      required: false
      schema:
        type: string
        format: date-time
    DateToFilter:
      name: dateTo
      in: query
      description: Fecha final inclusiva
      required: false
      schema:
        type: string
        format: date-time
    TransactionTypeIdFilter:
      name: transactionTypeId
      in: query
      description: Filtra por tipo de transaccion
      required: false
      schema:
        type: integer
        format: int64
    CategoryIdFilter:
      name: categoryId
      in: query
      description: Filtra por categoria principal
      required: false
      schema:
        type: integer
        format: int64
    TagFilter:
      name: tag
      in: query
      description: |
        Devuelve unicamente las transacciones que llevan el tag indicado.
        La comparacion no distingue mayusculas de minusculas.
      required: false
      schema:
        type: string
        maxLength: 70
  responses:
    BadRequest:
      description: La peticion es sintacticamente invalida
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: El recurso solicitado no existe
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: La operacion entra en conflicto con el estado actual del recurso
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    TransactionTypeResponse:
      type: object
      required: [id, name, code]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          description: Nombre para mostrar, por ejemplo Ingreso
        code:
          type: string
          description: |
            Codigo estable del tipo. Determina el signo del importe en los calculos.
          enum: [INCOME, EXPENSE]
    CategoryScope:
      type: string
      description: |
        Tipo de movimiento al que se ofrece una categoria. BOTH la ofrece en los dos.
        Solo decide que categorias propone el formulario de registro; no restringe lo que
        ya esta guardado.
      enum: [EXPENSE, INCOME, BOTH]
    CategoryResponse:
      type: object
      description: Categoria del catalogo del usuario
      required: [id, name, appliesTo, isSystem, transactionCount]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        appliesTo:
          $ref: "#/components/schemas/CategoryScope"
        isSystem:
          type: boolean
          description: |
            Categoria de reserva del usuario. Recibe las transacciones de las categorias
            que se eliminan y por eso no puede eliminarse ella misma.
        transactionCount:
          type: integer
          format: int64
          description: Cuantas transacciones del usuario clasifica esta categoria
    SaveCategoryRequest:
      type: object
      description: Cuerpo de alta y de actualizacion de una categoria
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 70
        appliesTo:
          $ref: "#/components/schemas/CategoryScope"
    TransactionResponse:
      type: object
      required: [id, amount, date, transactionType, category, tags]
      properties:
        id:
          type: integer
          format: int64
        amount:
          type: number
          description: Importe monetario con dos decimales, siempre positivo
        description:
          type: string
        date:
          type: string
          format: date-time
        transactionType:
          $ref: "#/components/schemas/TransactionTypeResponse"
        category:
          $ref: "#/components/schemas/CategoryResponse"
        tags:
          type: array
          description: Tags de la transaccion, en orden alfabetico
          items:
            type: string
    CreateTransactionRequest:
      type: object
      required: [amount, transactionTypeId, categoryId]
      properties:
        amount:
          type: number
          minimum: 0
          exclusiveMinimum: true
        description:
          type: string
          maxLength: 255
        date:
          type: string
          format: date-time
          description: Fecha de la transaccion. Si se omite se usa la fecha actual.
        transactionTypeId:
          type: integer
          format: int64
        categoryId:
          type: integer
          format: int64
          description: |
            Categoria principal, obligatoria. Debe pertenecer al usuario y admitir el tipo
            de la transaccion.
        tags:
          type: array
          description: |
            Tags de texto libre. Se recortan los espacios sobrantes, se descartan los que
            quedan en blanco y los repetidos se ignoran sin distinguir mayusculas.
            Una cadena vacia no llega a descartarse: la rechaza la longitud minima.
            Si el usuario ya tiene el tag escrito de otra forma se reutiliza el existente,
            asi que `casa` sobre un `Casa` previo devuelve `Casa`.
          items:
            type: string
            minLength: 1
            maxLength: 70
    UpdateTransactionRequest:
      type: object
      properties:
        amount:
          type: number
          minimum: 0
          exclusiveMinimum: true
        description:
          type: string
          maxLength: 255
        date:
          type: string
          format: date-time
        transactionTypeId:
          type: integer
          format: int64
        categoryId:
          type: integer
          format: int64
          description: |
            Cambia la categoria principal. Debe pertenecer al usuario y admitir el tipo
            que tendra la transaccion despues de la actualizacion.
        tags:
          type: array
          description: |
            Si esta presente reemplaza por completo los tags de la transaccion.
            Una lista vacia deja la transaccion sin tags.
          items:
            type: string
            minLength: 1
            maxLength: 70
    TransactionPageResponse:
      type: object
      description: Pagina de transacciones con sus metadatos de paginacion
      required: [content, page, size, totalElements, totalPages]
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/TransactionResponse"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
    TagResponse:
      type: object
      description: Tag del catalogo del usuario
      required: [id, name, transactionCount]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        transactionCount:
          type: integer
          format: int64
          description: Cuantas transacciones del usuario llevan este tag
    SaveTagRequest:
      type: object
      description: Nombre de un tag al crearlo o al renombrarlo
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 70
          description: |
            Se recortan los espacios sobrantes. No puede coincidir con otro tag del mismo
            usuario sin distinguir mayusculas.
...
    ErrorResponse:
      type: object
      description: Respuesta de error estructurada
      required: [timestamp, status, code, message]
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        code:
          type: string
          description: Codigo estable de negocio, por ejemplo TRANSACTION_NOT_FOUND
        message:
          type: string